+++ backend/config/test.exs
@@ -59,7 +59,8 @@ config :edgehog, :astarte_system_status_module, Edgehog.Astarte.Device.SystemSta
 config :edgehog, :astarte_battery_status_module, Edgehog.Astarte.Device.BatteryStatusMock
 config :edgehog, :astarte_base_image_module, Edgehog.Astarte.Device.BaseImageMock
 config :edgehog, :astarte_os_info_module, Edgehog.Astarte.Device.OSInfoMock
-config :edgehog, :astarte_ota_request_module, Edgehog.Astarte.Device.OTARequestMock
+config :edgehog, :astarte_ota_request_v0_module, Edgehog.Astarte.Device.OTARequestV0Mock
+config :edgehog, :astarte_ota_request_v1_module, Edgehog.Astarte.Device.OTARequestV1Mock
 config :edgehog, :astarte_runtime_info_module, Edgehog.Astarte.Device.RuntimeInfoMock
 config :edgehog, :astarte_led_behavior_module, Edgehog.Astarte.Device.LedBehaviorMock
 config :edgehog, :astarte_geolocation_module, Edgehog.Astarte.Device.GeolocationMock