+++ backend/test/edgehog/astarte_test.exs
@@ -1,7 +1,7 @@
 #
 # This file is part of Edgehog.
 #
-# Copyright 2021-2022 SECO Mind Srl
+# Copyright 2021-2023 SECO Mind Srl
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -22,12 +22,17 @@ defmodule Edgehog.AstarteTest do
   use Edgehog.DataCase
   use Edgehog.AstarteMockCase
 
+  alias Astarte.Client.APIError
+  alias Astarte.Client.AppEngine
   alias Edgehog.Astarte
+  alias Edgehog.Astarte.Device.DeviceStatus
+  alias Edgehog.Astarte.InterfaceVersion
+
+  import Edgehog.AstarteFixtures
+  import Edgehog.DevicesFixtures
 
   describe "clusters" do
     alias Edgehog.Astarte.Cluster
-
-    import Edgehog.AstarteFixtures
 
     @invalid_attrs %{base_api_url: nil, name: nil}
 
@@ -82,8 +87,6 @@ defmodule Edgehog.AstarteTest do
 
   describe "realms" do
     alias Edgehog.Astarte.Realm
-
-    import Edgehog.AstarteFixtures
 
     setup do
       %{cluster: cluster_fixture()}
@@ -142,9 +145,6 @@ defmodule Edgehog.AstarteTest do
 
   describe "devices" do
     alias Edgehog.Astarte.Device
-
-    import Edgehog.AstarteFixtures
-    import Edgehog.DevicesFixtures
 
     setup do
       cluster = cluster_fixture()
@@ -312,5 +312,212 @@ defmodule Edgehog.AstarteTest do
 
       assert device.part_number == part_number
     end
+  end
+
+  describe "send_ota_request_update/4" do
+    setup do
+      cluster = cluster_fixture()
+      realm = realm_fixture(cluster)
+
+      %{base_api_url: base_api_url} = cluster
+      %{name: realm_name, private_key: private_key} = realm
+
+      {:ok, client} = AppEngine.new(base_api_url, realm_name, private_key: private_key)
+
+      {:ok, device: device_fixture(realm), client: client}
+    end
+
+    test "sends the request for io.edgehog.devicemanager.OTARequest v1.0", ctx do
+      %{device: device, client: client} = ctx
+
+      introspection = %{
+        "io.edgehog.devicemanager.OTARequest" => %{major: 1, minor: 0}
+      }
+
+      mock_device_status_introspection(device, introspection)
+
+      device_id = device.device_id
+      uuid = Ecto.UUID.generate()
+      url = "https://mybucket.foo/update.bin"
+
+      Edgehog.Astarte.Device.OTARequestV1Mock
+      |> expect(:update, fn ^client, ^device_id, ^uuid, ^url ->
+        :ok
+      end)
+
+      assert :ok = Astarte.send_ota_request_update(client, device_id, uuid, url)
+    end
+
+    test "sends the request for io.edgehog.devicemanager.OTARequest v0.1", ctx do
+      %{device: device, client: client} = ctx
+
+      introspection = %{
+        "io.edgehog.devicemanager.OTARequest" => %{major: 0, minor: 1}
+      }
+
+      mock_device_status_introspection(device, introspection)
+
+      device_id = device.device_id
+      uuid = Ecto.UUID.generate()
+      url = "https://mybucket.foo/update.bin"
+
+      Edgehog.Astarte.Device.OTARequestV0Mock
+      |> expect(:post, fn ^client, ^device_id, ^uuid, ^url ->
+        :ok
+      end)
+
+      assert :ok = Astarte.send_ota_request_update(client, device_id, uuid, url)
+    end
+
+    test "fails if the API returns a failure", ctx do
+      %{device: device, client: client} = ctx
+
+      introspection = %{
+        "io.edgehog.devicemanager.OTARequest" => %{major: 1, minor: 0}
+      }
+
+      mock_device_status_introspection(device, introspection)
+
+      device_id = device.device_id
+      uuid = Ecto.UUID.generate()
+      url = "https://mybucket.foo/update.bin"
+
+      Edgehog.Astarte.Device.OTARequestV1Mock
+      |> expect(:update, fn ^client, ^device_id, ^uuid, ^url ->
+        {:error,
+         %APIError{
+           status: 500,
+           response: %{"errors" => %{"detail" => "Internal server error"}}
+         }}
+      end)
+
+      assert {:error, %APIError{status: 500}} =
+               Astarte.send_ota_request_update(client, device_id, uuid, url)
+    end
+
+    test "fails if the device doesn't have the OTARequest interface", ctx do
+      %{device: device, client: client} = ctx
+
+      introspection = []
+      mock_device_status_introspection(device, introspection)
+
+      device_id = device.device_id
+      uuid = Ecto.UUID.generate()
+      url = "https://mybucket.foo/update.bin"
+
+      Edgehog.Astarte.Device.OTARequestV0Mock
+      |> expect(:post, 0, fn _client, _device_id, _uuid, _url -> :ok end)
+
+      Edgehog.Astarte.Device.OTARequestV1Mock
+      |> expect(:update, 0, fn _client, _device_id, _uuid, _url -> :ok end)
+
+      assert {:error, :ota_request_not_supported} =
+               Astarte.send_ota_request_update(client, device_id, uuid, url)
+    end
+  end
+
+  describe "send_ota_request_cancel/3" do
+    setup do
+      cluster = cluster_fixture()
+      realm = realm_fixture(cluster)
+
+      %{base_api_url: base_api_url} = cluster
+      %{name: realm_name, private_key: private_key} = realm
+
+      {:ok, client} = AppEngine.new(base_api_url, realm_name, private_key: private_key)
+
+      {:ok, device: device_fixture(realm), client: client}
+    end
+
+    test "sends the request for io.edgehog.devicemanager.OTARequest v1.0", ctx do
+      %{device: device, client: client} = ctx
+
+      introspection = %{
+        "io.edgehog.devicemanager.OTARequest" => %{major: 1, minor: 0}
+      }
+
+      mock_device_status_introspection(device, introspection)
+
+      device_id = device.device_id
+      uuid = Ecto.UUID.generate()
+
+      Edgehog.Astarte.Device.OTARequestV1Mock
+      |> expect(:cancel, fn ^client, ^device_id, ^uuid ->
+        :ok
+      end)
+
+      assert :ok = Astarte.send_ota_request_cancel(client, device_id, uuid)
+    end
+
+    test "fails for io.edgehog.devicemanager.OTARequest v0.1", ctx do
+      %{device: device, client: client} = ctx
+
+      introspection = %{
+        "io.edgehog.devicemanager.OTARequest" => %{major: 0, minor: 1}
+      }
+
+      mock_device_status_introspection(device, introspection)
+
+      device_id = device.device_id
+      uuid = Ecto.UUID.generate()
+
+      assert {:error, :cancel_not_supported} =
+               Astarte.send_ota_request_cancel(client, device_id, uuid)
+    end
+
+    test "fails if the API returns a failure", ctx do
+      %{device: device, client: client} = ctx
+
+      introspection = %{
+        "io.edgehog.devicemanager.OTARequest" => %{major: 1, minor: 0}
+      }
+
+      mock_device_status_introspection(device, introspection)
+
+      device_id = device.device_id
+      uuid = Ecto.UUID.generate()
+
+      Edgehog.Astarte.Device.OTARequestV1Mock
+      |> expect(:cancel, fn ^client, ^device_id, ^uuid ->
+        {:error,
+         %APIError{
+           status: 500,
+           response: %{"errors" => %{"detail" => "Internal server error"}}
+         }}
+      end)
+
+      assert {:error, %APIError{status: 500}} =
+               Astarte.send_ota_request_cancel(client, device_id, uuid)
+    end
+
+    test "fails if the device doesn't have the OTARequest interface", ctx do
+      %{device: device, client: client} = ctx
+
+      introspection = []
+      mock_device_status_introspection(device, introspection)
+
+      device_id = device.device_id
+      uuid = Ecto.UUID.generate()
+
+      Edgehog.Astarte.Device.OTARequestV1Mock
+      |> expect(:cancel, 0, fn _client, _device_id, _uuid -> :ok end)
+
+      assert {:error, :ota_request_not_supported} =
+               Astarte.send_ota_request_cancel(client, device_id, uuid)
+    end
+  end
+
+  defp mock_device_status_introspection(device, interfaces) do
+    device_id = device.device_id
+
+    introspection =
+      for {name, %{major: major, minor: minor}} <- interfaces, into: %{} do
+        {name, %InterfaceVersion{major: major, minor: minor}}
+      end
+
+    Edgehog.Astarte.Device.DeviceStatusMock
+    |> expect(:get, fn _appengine_client, ^device_id ->
+      {:ok, %DeviceStatus{introspection: introspection}}
+    end)
   end
 end